Start qocw2 test
Starting with Read benchmark, since testing lima show that all time is
spent on reading the image. The benchmark measures the time to open a
compressed and uncompressed qcow2 images without a backing file and read
the entire image. This matches the common use of this library in lima.

Example run:

    % go test -bench Read
    BenchmarkRead/qcow2-12        1   1957581083 ns/op   137.13 MB/s   4297616648 B/op	  98474 allocs/op
    BenchmarkRead/qcow2_zlib-12   1   5352411666 ns/op    50.15 MB/s   5034875288 B/op   278804 allocs/op

Signed-off-by: Nir Soffer <nsoffer@redhat.com>